Fixed an issue regarding encoded non-English group/chapter names when downloading an entire series.
When attempting to download a series, the html parser would fail
to parse certain rows correctly. The row containing invalid characters
wouldn't be closed correctly, thus causing it to merge with the next row.
This would continue until a valid row, usually in English, was found,
resulting in a bunch of other languages being merged into an English row.

As a result, an English row would be parsed, as desired, but there would
be hrefs from previous, invalid rows as well.

This has been fixed by changing the way xpath looks for chapters.